When replying and later displaying replies, fill the whole quote line.
The current behaviour when replying to a message is to quote only the
first 100 characters when the new message is created.

When the new message is sent and rendered, only 60 of those 100
characters are actually displayed in the quoted area.

Since we have the width to display approximately 320 characters at
composition time, and 180 characters at rendering time (assuming a
full-screen client on a 4k desktop), let's use it.