Enable FTS5 full-text searches.
Session uses SQLite FTS5 tables that allow full-text searching, but also
includes code that renders this type of advanced search inoperable.

This patch restores the functionality, allowing the use of prefix
queries (e.g. 'lin*'), Boolean operators (AND, OR and NOT), initial
token queries (e.g. '^Linux') and NEAR groups.

Operators must appear in upper case, so 'AND' is a Boolean operator, but
'and' is a search for the literal word 'and'.

Note that whitespace implies a Boolean AND, so 'linux AND applications'
is syntactically equivalent to 'linux applications', either of which
searches for any message containing both words in any order, and not
necessarily appearing contiguously.

To search for the contiguous expression 'linux applications', surround
it with double quotes: '"linux applications"'.

Parentheses may be used to change operator precedence and build complex
expressions.

The BNF for the FTS query syntax is this:

```
<phrase>    := string [*]
<phrase>    := <phrase> + <phrase>
<neargroup> := NEAR ( <phrase> <phrase> ... [, N] )
<query>     := [ [-] <colspec> :] [^] <phrase>
<query>     := [ [-] <colspec> :] <neargroup>
<query>     := [ [-] <colspec> :] ( <query> )
<query>     := <query> AND <query>
<query>     := <query> OR <query>
<query>     := <query> NOT <query>
<colspec>   := colname
<colspec>   := { colname1 colname2 ... }
```

For more information, please see:

https://www.sqlite.org/fts5.html#full_text_query_syntax